% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMScanSGWH_SAR.R
\name{LMScanSGWH_SAR}
\alias{LMScanSGWH_SAR}
\title{LM Scan test of SGWH in SAR Models}
\usage{
LMScanSGWH_SAR(y = y, X = XX, W = W, mC = mC, it2 = it2, NNN = NNN)
}
\arguments{
\item{y}{: Data vector Rx1}

\item{X}{: Data vector Rxk include the constant}

\item{W}{: W matrix RxR}

\item{mC}{: minimúm cluster size.}

\item{it2}{: The number of Monte Carlo replications, e.g., 999, 9999}

\item{NNN}{: Set of windows (circular or elliptic).}
}
\value{
The value of the test and Most Likelihood Cluster.
}
\description{
Get the LM Scan Test for SGWH
}
\examples{

R=100
y<-runif(R,0,1)
XX<-as.matrix(cbind(matrix(1,R,1),runif(R,0,1)))
cx <- runif(R, 0, 1)
cy <- runif(R, 0, 1)
co <- cbind(cx,cy)
W <- spdep::nb2mat(spdep::knn2nb(spdep::knearneigh(co,k=5,longlat=F)))
mC <- 10
NN <- ScanSGWH::nn_ellipse(Cx=co[,1],Cy=co[,2],nn=50,p=30)
## Circles
NNN<-NN$circles
it2 <- 999
LMScan <- LMScanSGWH_SAR(y=y,X=XX,W=W,mC=mC,it2=it2,NNN=NNN)
## Ellipses
NNN<-NN$ellipses
it2 <- 100
LMScan <- LMScanSGWH_SAR(y=y,X=XX,W=W,mC=mC,it2=it2,NNN=NNN)
}
\references{
Chasco C., Le Gallo, J and López F.A. A Scan test for spatial groupwise heteroscedasticity in cross-sectional models with an application on houses prices in Madrid
\cr
}
\seealso{
\code{\link{nn_ellipse}}
}
