% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_ellipse.R
\name{nn_ellipse}
\alias{nn_ellipse}
\title{obtain circular or ellipetic windows for a set of coordianates}
\usage{
nn_ellipse(Cx = Cx, Cy = Cy, nn = nn, p = p)
}
\arguments{
\item{Cx}{: Coordenate (Longitude). A vector Rx1}

\item{Cy}{: Coordenate (Latitude). A vector Rx1}

\item{nn}{: maximun (number of observations) size of the windows.}

\item{p}{: rotation angle ellipse.}
}
\value{
ventanas \code{x} and \code{y}.
}
\description{
Obtain circular or ellipetic windows for a set of coordianates
}
\examples{

R <- 1000
nn<-50
Cx <- runif(R,0,1)
Cy <- runif(R,0,1)
A <- nn_ellipse(Cx=Cx,Cy=Cy,nn=nn,p=30)

}
