% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScanSGWH_v0.R
\name{ScanSGWH_v0}
\alias{ScanSGWH_v0}
\title{Scan Spatial Groupwise Heteroscedasticity (SGWH) for Normal process}
\usage{
ScanSGWH_v0(y = y, C = C, nv = nv, k = k, Nrep = Nrep)
}
\arguments{
\item{y}{: Data vector Rx1}

\item{C}{: Coordenate vector Rx2.}

\item{nv}{: maximun number of neighbors inside cluster.}

\item{k}{: Looking for only high incidence cluster (k=0); only low (k=1); both (k=2).}

\item{Nrep}{: The number of Monte Carlo replications, e.g., 999, 9999}
}
\value{
The value of the test and Most Likelihood Cluster.
}
\description{
Get the Scan Test for SGWH
}
\examples{

y <- runif(100,0,1)
nv <- 25
k <- 2
Nrep <- 10
C <- cbind(runif(100,0,1),runif(100,0,1))
LR<-Scan <- ScanSGWH_v0(y=y,C=C,nv=nv,k=k,Nrep=Nrep)
}
\references{
Chasco C., Le Gallo, J and López F.A. A Scan test for spatial groupwise heteroscedasticity in cross-sectional models with an application on houses prices in Madrid
}
